Fix stray int cast

+++ com-mmxlabs-models-lng-transformer/com.mmxlabs.models.lng.analytics.transformer/src/com/mmxlabs/models/lng/analytics/transformer/impl/AnalyticsTransformer.java
@@ -85,7 +85,6 @@ import com.mmxlabs.scheduler.optimiser.fitness.impl.SchedulerUtils;
 import com.mmxlabs.scheduler.optimiser.fitness.impl.VoyagePlanOptimiser;
 import com.mmxlabs.scheduler.optimiser.manipulators.SequencesManipulatorUtil;
 import com.mmxlabs.scheduler.optimiser.providers.IPortSlotProvider;
-import com.mmxlabs.scheduler.optimiser.providers.IRouteCostProvider;
 import com.mmxlabs.scheduler.optimiser.providers.IStartEndRequirementProvider;
 import com.mmxlabs.scheduler.optimiser.providers.IVesselProvider;
 import com.mmxlabs.scheduler.optimiser.providers.guice.DataComponentProviderModule;
@@ -546,7 +545,7 @@ public class AnalyticsTransformer implements IAnalyticsTransformer {
 			if (consumption == 0) {
 				continue;
 			}
-			final int componentCost = OptimiserUnitConvertor.convertToExternalFixedCost((int) Calculator.costFromConsumption(consumption, unitPrice));
+			final int componentCost = OptimiserUnitConvertor.convertToExternalFixedCost(Calculator.costFromConsumption(consumption, unitPrice));
 			totalFuelCost += componentCost;
 			final ExtraData componentData = fuelData.addExtraData(component.name(), component.name(), componentCost, ExtraDataFormatType.CURRENCY);
 			componentData.addExtraData("quantity", "Usage (" + component.getDefaultFuelUnit().name() + ")", OptimiserUnitConvertor.convertToExternalVolume(consumption), ExtraDataFormatType.INTEGER);
